Started implementing Multiple Levels
Added enum to GlobalEnums to represent the current level.

Created LevelData class to store level data.
    LevelData:
        -CurrentLevel               Enum represent
        -NumberOfEnemiesPerLevel    Dictionary to store number of enemies per
                                    level. Key is the level enum, value is the
                                    number of enemies by type (index 0: Level1Enemies
                                                               index 1: Level2Enemies
                                                               index 2: Level3Enemies
                                                               index 3: Level4Enemies).
        Has Method to move to the next level

Made changes to EnemyManager to use LevelData to get the number of enemies per
type to spawn for the current level.

Made changes to GameManager to update the current level in LevelData when the
player completes a level.